feat: handle discord dependency via Gradle & remove sources generation
Gradle will now handle downloading the Discord dependency as an APK and will use the artifact transformer registered by this plugin to convert it to a jar via dex2jar.

I enabled code generation for dex2jar, which means that the built-in decompiler in Android Studio will be sufficient to view the source code. As such, I removed explicit sources jar generation via JADX.